refactor(desktop): split personas.rs to clear file-size + Windows CI
personas.rs (2018 lines) was over the desktop file-size cap (carve-out
of 1279). Split the write-back code (find_team_for_persona_source,
write_back_persona_md, rewrite_persona_md + its tests) into
commands/personas/writeback.rs and the inbound-reconcile tests into
commands/personas/inbound_tests.rs. Pure relocation: the public command
surface is identical (all 8 #[tauri::command] fns unchanged, update_persona
call site unchanged); write_back_persona_md is now pub(super) so the parent
module can call it. Removes the personas.rs size carve-out — no file needs it.

Also fix the Windows-only failure in test_writeback_uses_manifest_path_not_convention:
assert on path components (file_name / parent file_name) instead of a
string .contains("personas/..."), which fails on Windows' `\` separator.
Test-only; the production write-back path uses source_path as a PathBuf and
is already cross-platform.
